Merging last actual branch to master

See merge request vsychev/boxingclub!26

+++ BoxingClub.DAL/Implementation/BoxingGroupRepository.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using Microsoft.EntityFrameworkCore;
 using System.Threading.Tasks;
 using System.Linq;
+using ArgumentNullException = BoxingClub.Infrastructure.Exceptions.ArgumentNullException;
 
 namespace BoxingClub.DAL.Repositories
 {
@@ -14,45 +15,82 @@ namespace BoxingClub.DAL.Repositories
 
         public BoxingGroupRepository(BoxingClubContext context)
         {
-            _db = context;
+            _db = context ?? throw new ArgumentNullException(nameof(context), "context is null");
         }
 
         public async Task CreateAsync(BoxingGroup item)
         {
+            if (item == null)
+            {
+                throw new ArgumentNullException(nameof(item), "BoxingGroup is null");
+            }
             await _db.BoxingGroups.AddAsync(item);
         }
 
         public void Delete(BoxingGroup item)
         {
+            if (item == null)
+            {
+                throw new ArgumentNullException(nameof(item), "BoxingGroup is null");
+            }
             _db.BoxingGroups.Remove(item);
         }
 
-        public async Task<BoxingGroup> GetByIdAsync(int id)
+        public Task<BoxingGroup> GetByIdAsync(int id)
         {
-            var res = await _db.BoxingGroups.AsQueryable().Where(g => g.Id == id).Include(x => x.Coach).SingleOrDefaultAsync();
-            return res;
+            return _db.BoxingGroups.Include(x => x.Coach).SingleOrDefaultAsync(g => g.Id == id);
         }
 
-        public async Task<IEnumerable<BoxingGroup>> GetAllAsync()
+        public Task<List<BoxingGroup>> GetAllAsync()
         {
-            return await _db.BoxingGroups.Include(x => x.Coach).ToListAsync();
+            return _db.BoxingGroups.Include(x => x.Coach).ToListAsync();
         }
 
-        public async Task<BoxingGroup> GetBoxingGroupWithStudentsByIdAsync(int id)
+        public Task<BoxingGroup> GetBoxingGroupWithStudentsByIdAsync(int id)
         {
-            var res = await _db.BoxingGroups.AsQueryable().Where(x => x.Id == id).Include(x => x.Coach).Include(x => x.Students).SingleOrDefaultAsync();
-            return res;
+           return _db.BoxingGroups.Include(x => x.Coach).Include(x => x.Students).SingleOrDefaultAsync(x => x.Id == id);
         }
 
         public void Update(BoxingGroup item)
         {
+            if (item == null)
+            {
+                throw new ArgumentNullException(nameof(item), "BoxingGroup is null");
+            }
             _db.Entry(item).State = EntityState.Modified;
         }
 
-        public async Task<List<BoxingGroup>> GetBoxingGroupsByCoachIdAsync(string id)
+        public Task<List<BoxingGroup>> GetBoxingGroupsByCoachIdAsync(string id)
         {
-            var groups = await _db.BoxingGroups.AsQueryable().Where(x => x.CoachId == id).ToListAsync();
-            return groups;
+            return _db.BoxingGroups.AsQueryable().Where(x => x.CoachId == id).ToListAsync();
+        }
+
+        public Task<List<BoxingGroup>> GetBoxingGroupsPaginatedAsync(int pageIndex, int pageSize)
+        {
+            var query = _db.BoxingGroups.Include(x => x.Coach);
+            var list = query.Skip((pageIndex - 1) * pageSize).Take(pageSize).ToListAsync();
+            return list;
+        }
+
+        public Task<int> GetCountOfBoxingGroupsAsync()
+        {
+            var query = _db.BoxingGroups.AsQueryable().Include(x => x.Coach);
+            var count = query.CountAsync();
+            return count;
+        }
+
+        public Task<List<BoxingGroup>> GetBoxingGroupsByCoachIdPaginatedAsync(string id, int pageIndex, int pageSize)
+        {
+            var query = _db.BoxingGroups.AsQueryable().Where(x => x.CoachId == id);
+            var list = query.Skip((pageIndex - 1) * pageSize).Take(pageSize).ToListAsync();
+            return list;
+        }
+
+        public Task<int> GetCountOfBoxingGroupsByCoachIdAsync(string id)
+        {
+            var query = _db.BoxingGroups.AsQueryable().Where(x => x.CoachId == id);
+            var count = query.CountAsync();
+            return count;
         }
     }
 }
